+++ module/auth/src/main/kotlin/io/github/siyual_park/auth/domain/authentication/AuthenticateFilterFactory.kt
@@ -1,4 +1,4 @@
-package io.github.siyual_park.auth.domain.authenticator
+package io.github.siyual_park.auth.domain.authentication
 
 import org.springframework.beans.BeansException
 import org.springframework.context.ApplicationContext
