+++ module/auth/src/main/kotlin/io/github/siyual_park/auth/domain/authentication/AuthenticateMapping.kt
@@ -1,4 +1,4 @@
-package io.github.siyual_park.auth.domain.authenticator
+package io.github.siyual_park.auth.domain.authentication
 
 import kotlin.reflect.KClass
