+++ module/auth/src/main/kotlin/io/github/siyual_park/auth/domain/authentication/AuthorizationPayload.kt
@@ -1,4 +1,4 @@
-package io.github.siyual_park.auth.domain.authenticator
+package io.github.siyual_park.auth.domain.authentication
 
 data class AuthorizationPayload(
     val type: String,
